Implement "forward energy" variant in new branch
Branching off from the original implementation because enough parts of
the algorithm have to change. In the future, I may want to find a way to
support both the older "backward energy" and the newer "forward energy"
variants in the same branch with the ability to switch between the two
at runtime.